--- lizardwizards/src/main/java/com/lizardwizards/lizardwizards/core/gameplay/CollisionShape.java
@@ -1,6 +1,0 @@
-package com.lizardwizards.lizardwizards.core.gameplay;
-
-public enum CollisionShape {
-    Circle,
-    Rectangle
-}
