+++ lizardwizards/src/main/java/com/lizardwizards/lizardwizards/core/gameplay/EntityWrapper.java
@@ -2,6 +2,8 @@ package com.lizardwizards.lizardwizards.core.gameplay;
 
 import com.lizardwizards.lizardwizards.client.EntitySprite;
 import com.lizardwizards.lizardwizards.core.Vector2;
+import com.lizardwizards.lizardwizards.core.gameplay.collision.Collider;
+import com.lizardwizards.lizardwizards.core.gameplay.collision.CollisionLayer;
 
 import java.io.Serializable;
 import java.util.HashMap;
