+++ lizardwizards/src/main/java/com/lizardwizards/lizardwizards/core/gameplay/collision/CollisionLayer.java
@@ -1,4 +1,4 @@
-package com.lizardwizards.lizardwizards.core.gameplay;
+package com.lizardwizards.lizardwizards.core.gameplay.collision;
 
 public enum CollisionLayer {
     Player,
