+++ lizardwizards/src/main/java/com/lizardwizards/lizardwizards/core/gameplay/collision/CollisionShape.java
@@ -1,0 +1,6 @@
+package com.lizardwizards.lizardwizards.core.gameplay.collision;
+
+public enum CollisionShape {
+    Circle,
+    Rectangle
+}
