+++ lizardwizards/src/main/java/com/lizardwizards/lizardwizards/core/gameplay/levels/GenerationRoomData.java
@@ -1,4 +1,6 @@
-package com.lizardwizards.lizardwizards.core.gameplay;
+package com.lizardwizards.lizardwizards.core.gameplay.levels;
+
+import com.lizardwizards.lizardwizards.core.gameplay.levels.RoomData;
 
 public class GenerationRoomData {
     RoomData roomData;
