+++ lizardwizards/src/main/java/com/lizardwizards/lizardwizards/core/gameplay/levels/RoomEnumerator.java
@@ -1,0 +1,6 @@
+package com.lizardwizards.lizardwizards.core.gameplay.levels;
+
+public enum RoomEnumerator {
+    Start,
+    BasicRoom
+}
